Warning fixes
git-svn-id: https://deskvox.svn.sourceforge.net/svnroot/deskvox/trunk@1334 04231f92-3938-0410-9931-e931ac552b4f

+++ virvo/virvo/vvcudaimg.cpp
@@ -125,7 +125,7 @@ void vvCudaImg::map()
     size_t size;
     vvCuda::checkError(&ok, cudaGraphicsResourceGetMappedPointer((void**)&d_img, &size, _imgRes),
                        "vvCudaImg::map() - get PBO mapping");
-    assert(size == width*height*vvSoftImg::PIXEL_SIZE);
+    assert(size == static_cast<size_t>(width*height*vvSoftImg::PIXEL_SIZE));
   }
   else
   {
